+++ NEWS.md
@@ -2,6 +2,9 @@
 - `with_namespace()` function added, to run code with a modified object search
   path.
 
+- Add `with_tempfile()` and `local_tempfile()` functions to create temporary
+  files which are cleanup up afterwards. (#32)
+
 - Remove the `code` argument from `local_` functions (#50).
 
 # 2.0.0